Commit to Latest

+++ SpartanHotels/SpartanHotels.Repository.Core/Customer.cs
@@ -19,7 +19,7 @@ namespace SpartanHotels.Repository.Core
             this.Reservations = new HashSet<Reservation>();
         }
     
-        public int CustomerID { get; set; }
+        public long CustomerID { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
